don't set a trap for SIGINT (EXIT trap is still called)
also a SIGINT trap without an own exit will cause that the function
execution continues even if set -e is set if the function is called
in a conditional (while and so on). i.e. not only set -e is deactivated
but also a trap for SIGINT without explicit exit no longer exits