update placeholder with upstream content if no modification was made
We don't update placeholders even if the user did not modify
it, currently -- we basically treat all as locally modified. With this
change we want to enable that content is updated according to an
upstream update as long as the content was not modified locally
(regardless if that happened manually or via pull-hook.sh). Which also
means, if you modify content of a placeholder via pull-hook, then this
content is no longer updated.